IOS-3230: Fix first review comments
