+++ Tangem/Common/Extensions/Blockchain+.swift
@@ -207,6 +207,9 @@ extension Blockchain {
             .kava(testnet: false),
             .kaspa,
             .ravencoin(testnet: false),
+            .cosmos(testnet: false),
+            .terraV1,
+            .terraV2,
         ]
     }()
