+++ Tangem/UIComponents/TokenIcon/TokenIconInfo.swift
@@ -1,0 +1,16 @@
+//
+//  TokenIcon.swift
+//  Tangem
+//
+//  Created by Sergey Balashov on 17.11.2022.
+//  Copyright © 2022 Tangem AG. All rights reserved.
+//
+
+import Foundation
+import BlockchainSdk
+
+struct TokenIconInfo: Hashable {
+    let name: String
+    let blockchainIconName: String?
+    let imageURL: URL?
+}
