+++ Tangem/UIComponents/TokenIcon/TokenIconInfoBuilder.swift
@@ -1,22 +1,21 @@
 //
-//  TokenIcon.swift
+//  TokenIconInfoBuilder.swift
 //  Tangem
 //
-//  Created by Sergey Balashov on 17.11.2022.
-//  Copyright © 2022 Tangem AG. All rights reserved.
+//  Created by Andrew Son on 03/05/23.
+//  Copyright © 2023 Tangem AG. All rights reserved.
 //
 
 import Foundation
 import BlockchainSdk
 
-struct TokenIconInfo: Hashable {
-    let name: String
-
-    var blockchainIconName: String?
-    var imageURL: URL?
-
-    init(with type: Amount.AmountType, blockchain: Blockchain) {
+struct TokenIconInfoBuilder {
+    func build(for type: Amount.AmountType, in blockchain: Blockchain) -> TokenIconInfo {
         let id: String?
+        let name: String
+        var blockchainIconName: String?
+        var imageURL: URL?
+
         switch type {
         case .coin, .reserve:
             id = blockchain.id
@@ -31,5 +30,7 @@ struct TokenIconInfo: Hashable {
             imageURL = TokenIconURLBuilder(baseURL: CoinsResponse.baseURL)
                 .iconURL(id: id, size: .large)
         }
+
+        return .init(name: name, blockchainIconName: blockchainIconName, imageURL: imageURL)
     }
 }
